Move pkg/util/printers to staging and update refs
