+++ pkg/kubectl/cmd/get/customcolumn.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/cli-runtime/pkg/printers"
 	"k8s.io/client-go/util/jsonpath"
-	utilprinters "k8s.io/kubernetes/pkg/kubectl/util/printers"
+	utilprinters "k8s.io/kubectl/pkg/util/printers"
 )
 
 var jsonRegexp = regexp.MustCompile("^\\{\\.?([^{}]+)\\}$|^\\.?([^{}]+)$")
