+++ pkg/kubectl/metricsutil/metrics_printer.go
@@ -24,7 +24,7 @@ import (
 	"k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"k8s.io/client-go/kubernetes/scheme"
-	"k8s.io/kubernetes/pkg/kubectl/util/printers"
+	"k8s.io/kubectl/pkg/util/printers"
 	metricsapi "k8s.io/metrics/pkg/apis/metrics"
 )
